# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AuthResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/dto.UserInfo'
    type: object
  dto.BudgetResponse:
    properties:
      amount:
        description: Category   CategoryResponse `json:"category"`
        type: number
      category_id:
        type: integer
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: integer
      period:
        type: string
      remaining_amount:
        type: number
      spent_amount:
        type: number
      start_date:
        type: string
    type: object
  dto.BudgetsListResponse:
    properties:
      budgets:
        items:
          $ref: '#/definitions/dto.BudgetResponse'
        type: array
    type: object
...
  dto.CreateBudgetRequest:
    properties:
      amount:
        description: CategoryID uint       `json:"category_id" validate:"required"`
        example: 500
        type: number
      period:
        enum:
        - weekly
        - monthly
        - yearly
        example: monthly
        type: string
      user_id:
        type: integer
    required:
    - amount
    - period
    type: object
...
  dto.ErrorResponse:
    properties:
      details:
        additionalProperties:
          type: string
        type: object
      error:
        example: Validation failed
        type: string
      message:
        example: Email is required
        type: string
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        description: UserID   uint   `json:"id"`
        example: user@example.com
        type: string
      password:
        example: password123
        type: string
    required:
    - email
    - password
    type: object
  dto.RegisterRequest:
    properties:
      confirm_password:
        type: string
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        maxLength: 50
        minLength: 2
        type: string
      last_name:
        example: Doe
        maxLength: 50
        minLength: 2
        type: string
      password:
        example: password123
        maxLength: 100
        minLength: 8
        type: string
      time_of_registration:
        example: "2024-01-15T10:30:00Z"
        type: string
    required:
    - confirm_password
    - email
    - first_name
    - last_name
    - password
    type: object
  dto.UserInfo:
    properties:
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 1
        type: integer
      last_name:
        example: Doe
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Система управления личными финансами
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Finance API
  version: "1.0"
paths:
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Деактивация refresh токена и выход из системы
      produces:
      - application/json
      responses:
        "200":
          description: Успешный выход
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Токен не найден
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Выход из системы
      tags:
      - Authentication
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Аутентификация пользователя и получение JWT токенов
      parameters:
      - description: Данные для входа
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная аутентификация
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Неверные данные для входа
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Неверный email или пароль
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Вход в систему
      tags:
      - Authentication
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Создание нового аккаунта пользователя в системе
      parameters:
      - description: Данные для регистрации
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно зарегистрирован
          schema:
            $ref: '#/definitions/dto.UserInfo'
        "400":
          description: Ошибка валидации данных
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Пользователь с таким email уже существует
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Регистрация нового пользователя
      tags:
      - Authentication
...
  /categories/{category_id}/budgets:
    get:
      consumes:
      - application/json
      description: Получение всех бюджетов пользователя для указанной категории
      parameters:
      - description: ID категории
        in: path
        name: category_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список бюджетов
          schema:
            $ref: '#/definitions/dto.BudgetsListResponse'
        "400":
          description: Неверный ID категории
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение списка бюджетов
      tags:
      - Budgets
    post:
      consumes:
      - application/json
      description: Создание нового бюджета для указанной категории
      parameters:
      - description: ID категории
        in: path
        name: category_id
        required: true
        type: integer
      - description: Данные для создания бюджета
        in: body
        name: budget
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBudgetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Бюджет успешно создан
          schema:
            $ref: '#/definitions/dto.BudgetResponse'
        "400":
          description: Ошибка валидации данных
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создание нового бюджета
      tags:
      - Budgets
  /categories/{category_id}/budgets/{budget_id}:
    delete:
      consumes:
      - application/json
      description: Удаление конкретного бюджета пользователя
      parameters:
      - description: ID категории
        in: path
        name: category_id
        required: true
        type: integer
      - description: ID бюджета
        in: path
        name: budget_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Бюджет успешно удален
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный ID категории или бюджета
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление бюджета
      tags:
      - Budgets
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"